--- src/main/java/kr/co/springbootweeklymission/domain/member/api/response/MemberResDTO.java
@@ -1,22 +1,0 @@
-package kr.co.springbootweeklymission.domain.member.api.response;
-
-import kr.co.springbootweeklymission.domain.member.entity.model.MemberStatus;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class MemberResDTO {
-    @Getter
-    @Builder
-    @NoArgsConstructor(access = AccessLevel.PRIVATE)
-    @AllArgsConstructor(access = AccessLevel.PRIVATE)
-    public static class READ {
-        private UUID memberId;
-        private MemberStatus memberStatus;
-    }
-}
