--- src/main/java/kr/co/springbootweeklymission/domain/member/dao/MemberRepository.java
@@ -1,9 +1,0 @@
-package kr.co.springbootweeklymission.domain.member.dao;
-
-import kr.co.springbootweeklymission.domain.member.entity.Member;
-
-import java.util.List;
-
-public interface MemberRepository {
-    List<Member> findMembersByBlack();
-}
