--- src/main/java/kr/co/springbootweeklymission/domain/member/entity/Member.java
@@ -1,35 +1,0 @@
-package kr.co.springbootweeklymission.domain.member.entity;
-
-import kr.co.springbootweeklymission.domain.member.api.response.MemberResDTO;
-import kr.co.springbootweeklymission.domain.member.entity.model.MemberStatus;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@EqualsAndHashCode
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class Member {
-    private UUID memberId;
-    private MemberStatus memberStatus;
-
-    @Builder
-    private Member(UUID memberId,
-                   MemberStatus memberStatus) {
-        this.memberId = memberId;
-        this.memberStatus = memberStatus;
-    }
-
-    public static MemberResDTO.READ toMemberReadDto(Member member) {
-        return MemberResDTO.READ.builder()
-                .memberId(member.memberId)
-                .memberStatus(member.memberStatus)
-                .build();
-    }
-
-    public boolean isBlackMember() {
-        return this.memberStatus == MemberStatus.BLACK;
-    }
-}
