--- src/main/java/kr/co/springbootweeklymission/domain/voucher/dao/VoucherRepository.java
@@ -1,15 +1,0 @@
-package kr.co.springbootweeklymission.domain.voucher.dao;
-
-import kr.co.springbootweeklymission.domain.voucher.entity.Voucher;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
-
-public interface VoucherRepository {
-    Voucher save(Voucher voucher);
-
-    Optional<Voucher> findById(UUID voucherId);
-
-    List<Voucher> findAll();
-}
