+++ src/main/java/kr/co/springbootweeklymission/infrastructure/error/exception/FileIOException.java
@@ -1,0 +1,9 @@
+package kr.co.springbootweeklymission.infrastructure.error.exception;
+
+import kr.co.springbootweeklymission.infrastructure.error.model.ResponseStatus;
+
+public class FileIOException extends RuntimeException {
+    public FileIOException(ResponseStatus responseStatus) {
+        super(responseStatus.getMessage());
+    }
+}
