+++ src/main/java/kr/co/springbootweeklymission/infrastructure/error/exception/NotFoundException.java
@@ -1,0 +1,9 @@
+package kr.co.springbootweeklymission.infrastructure.error.exception;
+
+import kr.co.springbootweeklymission.infrastructure.error.model.ResponseStatus;
+
+public class NotFoundException extends RuntimeException {
+    public NotFoundException(ResponseStatus responseStatus) {
+        super(responseStatus.getMessage());
+    }
+}
