+++ src/main/java/kr/co/springbootweeklymission/member/domain/repository/MemberRepository.java
@@ -1,0 +1,9 @@
+package kr.co.springbootweeklymission.member.domain.repository;
+
+import kr.co.springbootweeklymission.member.domain.entity.Member;
+
+import java.util.List;
+
+public interface MemberRepository {
+    List<Member> findMembersByBlack();
+}
